Updated the electrical system
Man this was a tough push, but now all the electrical systems including:
batteries, generators, multimeters, and wiring.

Beware though, this system is a WIP and i need to make it a more simpler
API and it's not commented yet. Also the wire blocks are currently
invisble but they still work just the same.