Added... alot of stuff... in a long time
It's been quite awhile since the last commit, and i've done alot of
stuff, i probably wont put everything i did on this list since it was so
long ago also.

-added the first tests of chunk entities, they are in very early testing
but are still pretty awesome. To make one, place the rocket core block
and then build other blocks onto it. Right click the chunk enitity with
an empty hand to pilot it.

-Cleaned up the chunk provider code.

-Made preparations for better star rendering (still not there yet)

-Added a new test electrical block, just a test for a new electrical
system.

-Made planet stone and ore's the same block with different meta.

-Added a new registry system where blocks register themselves in their
constructors.

-Added an ItemSimpleBlock class which handles different unlocalized
names for different meta variant of blocks.